Read experimental Multi-Parameter Mapping (MPM) data.

The function reads data generated in Multimodal Parameter Mapping (MPM) experiments.

readMPMData(t1Files = NULL, pdFiles = NULL, mtFiles = NULL, maskFile = NULL,
            TR = NULL, TE = NULL, FA = NULL, wghts = NULL, verbose = TRUE)

Arguments

t1Files
Vector of filenames corresponding to T1 weighted images (in Nifti-Format) with varying TE

pdFiles
Vector of filenames corresponding to PD weighted images (in Nifti-Format) with varying TE

mtFiles
optional Vector of filenames corresponding to MT weighted images (in Nifti-Format) with varying TE

maskFile
optional filename for mask (in Nifti-Format)

TR
optional numeric TR vector, if omitted information is extracted from .nii files if possible

TE
optional numeric TE vector, if omitted information is extracted from .nii files if possible

FA
optional numeric FA (flip-angle) vector, if omitted information is extracted from .nii files if possible

wghts
optional weights for MPM data volumes. Only needed is volumes have different data variance, e.g., in case of averages of multiple acquisitions.

verbose
logical - provide information on progress

Value

List with components

ddata

mpm data

sdim

dimension of image cube

nFiles

number of images / image files

t1Files

character - filenames of t1Files

pdFiles

character - filenames of pdFiles

mtFiles

character - filenames of mtFiles

model

Number of the ESTATICS model that can be used

maskFile

character - filenames of maskFile

mask

mask

TR

vector of TR values

TE

vector of TE values

FA

vector of FA values

and class-attribute 'mpmData'
References

Weiskopf, N.; Suckling, J.; Williams, G.; Correia, M. M.; Inkster, B.; Tait, R.; Ooi, C.; Bullmore, E. T. & Lutti, A. Quantitative multi-parameter mapping of R1, PD(*), MT, and R2(*) at 3T: a multi-center validation. Front Neurosci, Wellcome Trust Centre for Neuroimaging, UCL Institute of Neurology, University College London, UK., 2013, 7, 95

J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Doi:10.1007/978-3-030-29184-6.

Author

Karsten Tabelow tabelow@wias-berlin.de
J\"org Polzehl polzehl@wias-berlin.de

See also

Examples

dataDir <- system.file("extdata",package="qMRI")
#
#  set file names for T1w, MTw and PDw images
#
t1Names <- paste0("t1w_",1:8,".nii.gz")
mtNames <- paste0("mtw_",1:6,".nii.gz")
pdNames <- paste0("pdw_",1:8,".nii.gz")
t1Files <- file.path(dataDir, t1Names)
mtFiles <- file.path(dataDir, mtNames)
pdFiles <- file.path(dataDir, pdNames)
#
#  file names of mask and B1 field map
#
B1File <- file.path(dataDir, "B1map.nii.gz")
maskFile <- file.path(dataDir, "mask.nii.gz")
#
#  Acquisition parameters (TE, TR, Flip Angle) for T1w, MTw and PDw images
#
TE <- c(2.3, 4.6, 6.9, 9.2, 11.5, 13.8, 16.1, 18.4,
        2.3, 4.6, 6.9, 9.2, 11.5, 13.8,
        2.3, 4.6, 6.9, 9.2, 11.5, 13.8, 16.1, 18.4)
TR <- rep(25, 22)
FA <- c(rep(21, 8), rep(6, 6), rep(6, 8))
#
#   read MPM example data
#
library(qMRI)
mpm <- readMPMData(t1Files, pdFiles, mtFiles,
                   maskFile, TR = TR, TE = TE,
                   FA = FA, verbose = FALSE)